// Mission-1/Mission-1-1/src/types/movie.ts
export type Movie = {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
};

export type MovieResponse = {
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
}

// 지원하는 언어 코드
export type MovieLanguage = 'ko-KR' | 'en-US' | 'ja-JP';

// 검색 필터 상태
export type MovieFilters = {
    query: string;
    includeAdult: boolean;
    language: MovieLanguage;
};